move everything to an emails concern
